Initial COVID-19 data processing
The initial processing of raw COVID-19 data at the county level was done using a script written for this purpose. Initial data, which pertained to a large number of countries and regions, was first filtered based on the FIPS number. The FIPS number consists of five digits that are unique for identifying counties and similar regions in the United States. Areas with FIPS numbers were used in further analysis. Data that were not defined as counties (data from unknown sampling locations, cruise ships) were also excluded. Raw data was sourced from the Johns Hopkins Coronavirus Resource Center website.