Determing_second_peak_and_merging_with_socio-demographic_data
The next step involves extracting data corresponding to the number of detected infection cases and the number of deaths during the second wave of the pandemic in the United States. A dedicated R script was used for this purpose. The script filtered cumulative data on infection and death cases during the duration of the second wave of the pandemic, with the start and end dates of the wave determined for each state and then applied to all counties in that state. Using the same R script, data on infection and death cases during the second wave of the pandemic were merged with socio-demographic data for the counties