Data_transformation
Among the collected variables, it was observed that some of them deviate from a normal distribution. As the methods used in later work are sensitive to outliers and require a normal distribution of data, it was necessary to apply appropriate data transformations before applying these methods to approximate a normal distribution. Using standard transformations, skewness in the distribution is brought as close to zero as possible, i.e., the distribution is approximated to a normal distribution where the mean and median are equal. For all transformed variables where there was a change in direction during the transformation, a minus sign was added to cancel out this change. The reason for this is to ensure that the original and transformed variables are oriented in the same direction for easier interpretation of results. After all variables were transformed, any remaining outliers were replaced with the median of the transformed data.