Lasso Regression
For each tested λ value, the average mean squared error was determined for the training and test datasets, calculated as the mean of the mean squared errors from all repetitions. The best model was selected as the one with the largest λ value among all models with MSE within one standard error of the minimum MSE obtained. As higher λ values lead to a reduction in the number of retained parameters, the selected model was the most restrictive among those with approximately the highest accuracy. The λ value of that model was used in the final LASSO regression, performed on the entire dataset. Data was graphically represented.