ElasticNet Regression
The same setup as described earlier for Lasso regression was used for cross-validation. For each of the 100 α values in the range from 0.01 to 1, with steps of 0.01, 100 λ values were determined, resulting in a total of 10 thousand models, from which the best were selected through cross-validation. The final model was chosen as the one within one standard error of the model with the minimum MSE, retaining the fewest predictors. The λ and α values of this model were used in the final Elastic Net regression (on the entire dataset). Data was graphically represented.