ElasticNet Relaxed Model
Relaxed models were used to reduce noise resulting from a large number of predictors. These models involve two rounds of regression. In the second round of regression, only the predictors selected in the final models of the first round are included. Apart from this difference in input data, the process of the second round of regression is identical to the first. Data was graphically represented.